Project Oxford SDK Release - December 2015 Update 2

+++ Face/Android/ClientLibrary/lib/src/main/java/com/microsoft/projectoxford/face/contract/SimilarFace.java
@@ -37,5 +37,5 @@ import java.util.UUID;
 public class SimilarFace {
     public UUID faceId;
 
-    public float confidence;
+    public double confidence;
 }
